pyproject.toml
[tool.poetry]
name = "atm-api"
version = "0.1.0"
description = "Atm simulation api"
authors = ["Viktor Ashcheulov"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.6"
django-heroku = "^0.3.1"
gunicorn = "^20.0.4"
flake8 = "^3.8.1"
black = "^19.10b0"
python-dotenv = "^0.13.0"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
